refactor: findWinner 메서드에 대한 책임 이동
